+++ pkg/providers/appsec/testdata/TestResWAFRuleset/waf_ruleset_create_with_invalid_attack_group.tf
@@ -1,0 +1,16 @@
+provider "akamai" {
+  edgerc        = "../../common/testutils/edgerc"
+  cache_enabled = false
+}
+
+resource "akamai_appsec_waf_ruleset" "test" {
+  config_id          = 111111
+  security_policy_id = "2222_333333"
+
+  attack_groups = [
+    {
+      attack_group        = "INVALID_GROUP"
+      attack_group_action = "alert"
+    }
+  ]
+}