+++ pkg/providers/appsec/testdata/TestResWAFRuleset/waf_ruleset_create_with_invalid_rule_id.tf
@@ -1,0 +1,16 @@
+provider "akamai" {
+  edgerc        = "../../common/testutils/edgerc"
+  cache_enabled = false
+}
+
+resource "akamai_appsec_waf_ruleset" "test" {
+  config_id          = 111111
+  security_policy_id = "2222_333333"
+
+  rules = [
+    {
+      rule_id     = 999999
+      rule_action = "alert"
+    }
+  ]
+}