+++ pkg/providers/appsec/testdata/TestResWAFRuleset/waf_ruleset_missing_security_policy_id.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc        = "../../common/testutils/edgerc"
+  cache_enabled = false
+}
+
+resource "akamai_appsec_waf_ruleset" "test" {
+  config_id = 111111
+}